Finish generalizing warnings
